api/requirements.txt
flask==0.12.2
python-dotenv==0.7.1
pyjwt==1.5.3
